fix: 修改统一响应包装类

+++ somta-core/src/main/java/net/somta/core/base/result/ResponseDataResult.java
@@ -34,7 +34,7 @@ public class ResponseDataResult<T> extends ResponseResult {
     	return setErrorResponseResult(errorCode,errorMessage, null);
     }
 
-    public static ResponseDataResult setErrorResponseResult(int errorCode ,String errorMessage,Object data) {
+    public static ResponseDataResult setErrorResponseResult(long errorCode ,String errorMessage,Object data) {
         ResponseDataResult r = new ResponseDataResult();
         r.setSuccess(false);
         r.setErrorCode(errorCode);
